Reorganize system props code, improve doc comments
