+++ lib/src/mui_suggestors/system_props.dart
@@ -1,0 +1,139 @@
+// Copyright 2026 Workiva Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import 'package:analyzer/dart/element/element.dart';
+
+/// Duck-types a component props class as a MUI-system-props-supporting props
+/// class by checking for the presence of an `sx` prop and at least one system prop.
+bool hasSxAndSomeSystemProps(InterfaceElement propsElement) {
+  const propsToCheck = [
+    'sx',
+    // The following items are arbitrary; we don't need to check for all system props,
+    // just a few to help prevent false positives where components have a prop or two
+    // that happens to match a system prop.
+    'bgcolor',
+    'm',
+    'letterSpacing',
+  ];
+
+  return propsToCheck.every((propName) =>
+      propsElement.lookUpGetter(propName, propsElement.library) != null);
+}
+
+/// The names of all the MUI system props.
+const systemPropNames = {
+  'm',
+  'mt',
+  'mr',
+  'mb',
+  'ml',
+  'mx',
+  'my',
+  'p',
+  'pt',
+  'pr',
+  'pb',
+  'pl',
+  'px',
+  'py',
+  'width',
+  'maxWidth',
+  'minWidth',
+  'height',
+  'maxHeight',
+  'minHeight',
+  'boxSizing',
+  'display',
+  'displayPrint',
+  'overflow',
+  'textOverflow',
+  'visibility',
+  'whiteSpace',
+  'flexBasis',
+  'flexDirection',
+  'flexWrap',
+  'justifyContent',
+  'alignItems',
+  'alignContent',
+  'order',
+  'flex',
+  'flexGrow',
+  'flexShrink',
+  'alignSelf',
+  'justifyItems',
+  'justifySelf',
+  'gap',
+  'columnGap',
+  'rowGap',
+  'gridColumn',
+  'gridRow',
+  'gridAutoFlow',
+  'gridAutoColumns',
+  'gridAutoRows',
+  'gridTemplateColumns',
+  'gridTemplateRows',
+  'gridTemplateAreas',
+  'gridArea',
+  'bgcolor',
+  'color',
+  'zIndex',
+  'position',
+  'top',
+  'right',
+  'bottom',
+  'left',
+  'boxShadow',
+  'border',
+  'borderTop',
+  'borderRight',
+  'borderBottom',
+  'borderLeft',
+  'borderColor',
+  'borderRadius',
+  'fontFamily',
+  'fontSize',
+  'fontStyle',
+  'fontWeight',
+  'letterSpacing',
+  'lineHeight',
+  'textAlign',
+  'textTransform',
+  'margin',
+  'marginTop',
+  'marginRight',
+  'marginBottom',
+  'marginLeft',
+  'marginX',
+  'marginY',
+  'marginInline',
+  'marginInlineStart',
+  'marginInlineEnd',
+  'marginBlock',
+  'marginBlockStart',
+  'marginBlockEnd',
+  'padding',
+  'paddingTop',
+  'paddingRight',
+  'paddingBottom',
+  'paddingLeft',
+  'paddingX',
+  'paddingY',
+  'paddingInline',
+  'paddingInlineStart',
+  'paddingInlineEnd',
+  'paddingBlock',
+  'paddingBlockStart',
+  'paddingBlockEnd',
+  'typography',
+};
